requirements.txt
python-dotenv
openai
GitPython
radon
bandit
requests
GitPython